requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==23.1.0
autopep8==2.0.2
billiard==3.6.4.0
blinker==1.6.2
Brotli==1.0.9
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-filter==23.2
django-redis==5.2.0
django-rest-framework-nested==0.0.1
django-silk
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-filters==0.10.2
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docopt==0.6.2
drf-nested-routers==0.93.4
exceptiongroup==1.1.1
fastapi==0.95.1
Flask==2.3.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.9
gprof2dot==2022.7.29
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
locust==2.15.1
MarkupSafe==2.1.2
model-bakery==1.11.0
msgpack==1.0.5
mysqlclient==2.1.1
nltk==3.8.1
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
product==0.1.1.1
prompt-toolkit==3.0.38
psutil==5.9.5
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pytest==7.3.1
pytest-django==4.5.2
pytest-watch==4.2.0
python3-openid==3.2.0
pytz==2023.3
pywin32
pyzmq==25.0.2
redis==4.5.4
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
roundrobin==0.0.4
router==0.1
routers==0.10.1
simplerouter==1.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
starlette==0.26.1
style==1.1.0
textblob==0.17.1
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
update==0.0.1
uritemplate==4.1.1
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
watchdog==3.0.0
wcwidth==0.2.6
WebOb==1.8.7
Werkzeug==2.3.3
whitenoise==6.4.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
